Moved apis with same names into the same directory
