+++ server/api/image/index.get.ts
