+++ server/api/image/index.put.ts
